refactor: complete dependency inversion — GlassTabBar is now the single source of truth
- Create shared/tab_bar_bottom_layout.dart (TabBarBottomLayout StatefulWidget)
  Owns all bottom-placement rendering logic extracted from _GlassBottomBarState.

- Create shared/tab_bar_searchable_layout.dart (TabBarSearchableLayout StatefulWidget)
  Owns all searchable-placement rendering + spring physics extracted from
  _GlassSearchableBottomBarState.

- glass_tab_bar.dart: remove bridge class calls, import layout shells directly
  _buildBottom → TabBarBottomLayout, _buildSearchable → TabBarSearchableLayout
  Delete _GlassBottomBarBridge and _GlassSearchableBridge (~310 lines removed)

- GlassBottomBar: StatefulWidget → StatelessWidget shim (~240 lines removed)
  Delegates entire build() to TabBarBottomLayout via a single forwarding call.

- GlassSearchableBottomBar: StatefulWidget → StatelessWidget shim (~740 lines removed)
  Delegates entire build() to TabBarSearchableLayout via a single forwarding call.

- Update dispatch tests to assert new architecture (no bridge widgets in tree)

Net result: one fix in a layout shell propagates to both GlassTabBar.*() and the
deprecated shims automatically. Deleting v1 shims in a future PR is now trivial.

flutter analyze: 0 errors | flutter test: 2234/2234 passing
