+++ lib/liquid_glass_widgets.dart
@@ -108,7 +108,7 @@ export 'widgets/surfaces/glass_scaffold.dart';
 export 'widgets/surfaces/glass_bottom_bar.dart';
 export 'widgets/surfaces/glass_searchable_bottom_bar.dart';
 export 'widgets/surfaces/shared/glass_search_bar_config.dart';
-export 'widgets/surfaces/shared/searchable_bottom_bar_controller.dart';
+export 'widgets/surfaces/shared/tab_bar_searchable_controller.dart';
 
 export 'widgets/surfaces/glass_tab_bar.dart';
 export 'widgets/surfaces/glass_toolbar.dart';
