+++ test/widgets/surfaces/adaptive_brightness_bar_test.dart
@@ -14,7 +14,7 @@ import 'package:flutter/foundation.dart' show ValueListenable;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
-import 'package:liquid_glass_widgets/widgets/surfaces/shared/bottom_bar_internal.dart'
+import 'package:liquid_glass_widgets/widgets/surfaces/shared/tab_bar_bottom_internal.dart'
     show resolveBarLabelColor;
 
 /// Records the ambient brightness its build context sees — used as a tab
