+++ test/widgets/surfaces/searchable_bottom_bar_internal_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
-import 'package:liquid_glass_widgets/widgets/surfaces/shared/searchable_bottom_bar_internal.dart';
+import 'package:liquid_glass_widgets/widgets/surfaces/shared/tab_bar_searchable_internal.dart';
 
 // ---------------------------------------------------------------------------
 // Shared helpers
